+++ frontend/src/components/inventory-ui/DefaultOptionsForm.tsx
@@ -1,4 +1,4 @@
-import { useFormContext } from "react-hook-form";
+import { useFormContext } from 'react-hook-form';
 import {
   FormControl,
   FormDescription,
@@ -6,38 +6,46 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from "@/components/ui/form"
+} from '@/components/ui/form';
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@/components/ui/select"
-import { Separator } from "@/components/ui/separator"
-import { ColorOption, TagOption } from "@/types/options";
-import * as imsService from '@/ims-service'
-import { useQuery } from "@tanstack/react-query";
-import { InventorySettingsFormData } from "@/schemas/InventorySettingsSchema";
+} from '@/components/ui/select';
+import { Separator } from '@/components/ui/separator';
+import { ColorOption, TagOption } from '@/types/options';
+import * as imsService from '@/ims-service';
+import { useQuery } from '@tanstack/react-query';
+import { InventorySettingsFormData } from '@/schemas/InventorySettingsSchema';
 import StatusMultiSelect from './StatusMultiSelect';
 
 const DefaultOptionsForm = () => {
   const { control } = useFormContext<InventorySettingsFormData>();
 
-  const { data: statusData, isLoading: isStatusDataLoading } = useQuery<ColorOption[]>({
+  const { data: statusData, isLoading: isStatusDataLoading } = useQuery<
+    ColorOption[]
+  >({
     queryKey: ['fetchOptionValues', 'status'],
     queryFn: () => imsService.fetchOptionValues('status'),
   });
 
-  const { data: categoryData, isLoading: isCategoryDataLoading } = useQuery<TagOption[]>({
+  const { data: categoryData, isLoading: isCategoryDataLoading } = useQuery<
+    TagOption[]
+  >({
     queryKey: ['fetchOptionValues', 'category'],
     queryFn: () => imsService.fetchOptionValues('category'),
   });
 
-  const { data: equipmentTypeData, isLoading: isEquipmentTypeDataLoading } = useQuery<string[]>({
-    queryKey: ['fetchOptionValues', 'equipmentType'],
-    queryFn: () => imsService.fetchOptionValues('equipmentType'),
-  });
+  const { data: equipmentTypeData, isLoading: isEquipmentTypeDataLoading } =
+    useQuery<string[]>({
+      queryKey: ['fetchOptionValues', 'equipmentType'],
+      queryFn: () => imsService.fetchOptionValues('equipmentType'),
+    });
+
+  const isFormDataLoading =
+    isStatusDataLoading || isCategoryDataLoading || isEquipmentTypeDataLoading;
 
   return (
     <div className="w-full flex flex-col">
@@ -54,11 +62,7 @@ const DefaultOptionsForm = () => {
           <FormItem className="pb-2">
             <FormLabel className="font-medium">Status</FormLabel>
             <div className="flex md:w-2/3 gap-1">
-              {!(
-                isStatusDataLoading ||
-                isCategoryDataLoading ||
-                isEquipmentTypeDataLoading
-              ) ? (
+              {!isFormDataLoading ? (
                 <Select
                   disabled={isStatusDataLoading}
                   onValueChange={field.onChange}
@@ -98,16 +102,12 @@ const DefaultOptionsForm = () => {
       />
       <FormField
         control={control}
-        name="category"
+        name="hardwareCategory"
         render={({ field }) => (
           <FormItem className="pb-2">
-            <FormLabel className="font-medium">Category</FormLabel>
+            <FormLabel className="font-medium">Hardware Category</FormLabel>
             <div className="flex md:w-2/3 gap-1">
-              {!(
-                isStatusDataLoading ||
-                isCategoryDataLoading ||
-                isEquipmentTypeDataLoading
-              ) ? (
+              {!isFormDataLoading ? (
                 <Select
                   disabled={isCategoryDataLoading}
                   onValueChange={field.onChange}
@@ -115,7 +115,7 @@ const DefaultOptionsForm = () => {
                 >
                   <FormControl>
                     <SelectTrigger>
-                      <SelectValue placeholder="Select default category" />
+                      <SelectValue placeholder="Select default hardware category" />
                     </SelectTrigger>
                   </FormControl>
                   <SelectContent>
@@ -126,15 +126,64 @@ const DefaultOptionsForm = () => {
                       -
                     </SelectItem>
                     {categoryData &&
-                      categoryData.map((category: TagOption) => (
-                        <SelectItem
-                          key={category.value}
-                          value={category.value}
-                          className="w-full"
-                        >
-                          {category.value}
-                        </SelectItem>
-                      ))}
+                      categoryData
+                        .filter((category) => category.type === 'Hardware')
+                        .map((category: TagOption) => (
+                          <SelectItem
+                            key={category.value}
+                            value={category.value}
+                            className="w-full"
+                          >
+                            {category.value}
+                          </SelectItem>
+                        ))}
+                  </SelectContent>
+                </Select>
+              ) : (
+                <div className="h-10 w-full border rounded-md" />
+              )}
+            </div>
+            <FormMessage />
+          </FormItem>
+        )}
+      />
+      <FormField
+        control={control}
+        name="softwareCategory"
+        render={({ field }) => (
+          <FormItem className="pb-2">
+            <FormLabel className="font-medium">Software Category</FormLabel>
+            <div className="flex md:w-2/3 gap-1">
+              {!isFormDataLoading ? (
+                <Select
+                  disabled={isCategoryDataLoading}
+                  onValueChange={field.onChange}
+                  defaultValue={field.value}
+                >
+                  <FormControl>
+                    <SelectTrigger>
+                      <SelectValue placeholder="Select default software category" />
+                    </SelectTrigger>
+                  </FormControl>
+                  <SelectContent>
+                    <SelectItem
+                      value={'-'}
+                      className="w-full text-accent-foreground"
+                    >
+                      -
+                    </SelectItem>
+                    {categoryData &&
+                      categoryData
+                        .filter((category) => category.type === 'Software')
+                        .map((category: TagOption) => (
+                          <SelectItem
+                            key={category.value}
+                            value={category.value}
+                            className="w-full"
+                          >
+                            {category.value}
+                          </SelectItem>
+                        ))}
                   </SelectContent>
                 </Select>
               ) : (
@@ -152,11 +201,7 @@ const DefaultOptionsForm = () => {
           <FormItem className="pb-2">
             <FormLabel className="font-medium">Equipment type</FormLabel>
             <div className="flex md:w-2/3 gap-1">
-              {!(
-                isStatusDataLoading ||
-                isCategoryDataLoading ||
-                isEquipmentTypeDataLoading
-              ) ? (
+              {!isFormDataLoading ? (
                 <Select
                   disabled={isEquipmentTypeDataLoading}
                   onValueChange={field.onChange}
@@ -196,29 +241,30 @@ const DefaultOptionsForm = () => {
       />
       <FormField
         control={control}
-        name='deployableStatus'
+        name="deployableStatus"
         render={({ field }) => (
-          <FormItem className='pb-2'>
-            <FormLabel className='font-medium'>
+          <FormItem className="pb-2">
+            <FormLabel className="font-medium">
               Statuses for deployable assets
             </FormLabel>
             <FormControl>
-            <div className="flex md:w-2/3 gap-1">
-              {statusData && (
-                <StatusMultiSelect
-                  type='deployable'
-                  options={statusData.filter(option=>option.value!=='Deployed')}
-                  selected={field.value}
-                  onChange={field.onChange}
-                  placeholder="Select statuses..."
-                />
-              )}
+              <div className="flex md:w-2/3 gap-1">
+                {statusData && (
+                  <StatusMultiSelect
+                    type="deployable"
+                    options={statusData.filter(
+                      (option) => option.value !== 'Deployed'
+                    )}
+                    selected={field.value}
+                    onChange={field.onChange}
+                    placeholder="Select statuses..."
+                  />
+                )}
               </div>
             </FormControl>
             <FormMessage />
             <FormDescription>
-              These statuses signify assets that can be
-              deployed.
+              These statuses signify assets that can be deployed.
             </FormDescription>
           </FormItem>
         )}
@@ -281,6 +327,6 @@ const DefaultOptionsForm = () => {
       <Separator className="my-4" />
     </div>
   );
-}
+};
 
 export default DefaultOptionsForm;